# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscription-plans/{id}:
    get:
      operationId: getSubscriptionPlan
      tags: [subscription-plans]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Subscription plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionPlan"

  # ─── Tenants ──────────────────────────────────────────────────
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (super admin)
      parameters:
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantList"
    post:
      operationId: createTenant
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}:
    get:
      operationId: getTenant
      tags: [tenants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}/suspend:
    post:
      operationId: suspendTenant
      tags: [tenants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Tenant suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{id}/activate:
    post:
      operationId: activateTenant
      tags: [tenants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Tenant activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ─── Restaurants ──────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      parameters:
        - { name: restaurantId, in: query, schema: { type: integer } }
        - { name: role, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ─── Floor Tables ─────────────────────────────────────────────
...
  /restaurants/{restaurantId}/staff:
    get:
      operationId: listStaff
      tags: [staff]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: role, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
...
components:
  schemas:
...
    SubscriptionPlan:
      type: object
      required: [id, name, slug, price, billingPeriod]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        price: { type: string }
        billingPeriod: { type: string }
        maxRestaurants: { type: integer }
        maxBranches: { type: integer }
        maxStaff: { type: integer }
        maxTables: { type: integer }
        maxMenuItems: { type: integer }
        trialDays: { type: integer }
        features:
          type: array
          items: { type: string }
        isActive: { type: boolean }
        createdAt: { type: string }

    # ── Tenants ──
    Tenant:
      type: object
      required: [id, name, slug, planStatus]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        planId: { type: ["integer", "null"] }
        planStatus: { type: string }
        trialEndsAt: { type: ["string", "null"] }
        isActive: { type: boolean }
        isSuspended: { type: boolean }
        logoUrl: { type: ["string", "null"] }
        primaryColor: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    TenantList:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Tenant"
        total: { type: integer }
    TenantInput:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        planId: { type: integer }
        primaryColor: { type: string }
    TenantUpdate:
      type: object
      properties:
        name: { type: string }
        planId: { type: integer }
        planStatus: { type: string }
        isActive: { type: boolean }
        logoUrl: { type: string }
        primaryColor: { type: string }

    # ── Restaurants ──
...
    User:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        restaurantId: { type: ["integer", "null"] }
        lastLoginAt: { type: ["string", "null"] }
        createdAt: { type: string }
    UserInput:
      type: object
      required: [name, email, role, passwordHash]
      properties:
        name: { type: string }
        email: { type: string }
        passwordHash: { type: string }
        role: { type: string }
        phone: { type: string }
        restaurantId: { type: integer }
        tenantId: { type: integer }
    UserUpdate:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        role: { type: string }
        isActive: { type: boolean }
        avatarUrl: { type: string }
    StaffMember:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        phone: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        lastLoginAt: { type: ["string", "null"] }
        createdAt: { type: string }

    # ── Floor Tables ──
...